Switch from repackaged-asm to the original asm
It seems that as of ASM 4, backwards compatibility is now a goal. Since
older versions of ASM cannot handle Java 7 byte code, and pojomatic
requires Java 7, its unlikely to be running in a project still using
ASM 3.

Making this change should help reduce the effective footprint of pojomatic,
both because use of repackaged ASM never really took hold in the community
(especially now that they preserve backwards compatibility), and because
we only need to use the core ASM library, where as repackaged ASM was
a repackaging of asm-all.

There was one class being used by pojomatic that lived in asm-utils as
opposed to asm, CheckClassAdapter. However, while this class can be useful
when debuging code generation issues, it doesn't really need to be used in production. Hence, its use has been removed.