Revert usages of visitMethodInsn and the Handle constructor to ASM 4 compatibility.
In ASM version 6, an overload for each of these was added with an additional boolean
parameter. For our use cases, the version without the parameter correctly forwards
to the one with with the correct argument. By relying on the deprecated version,
we can preserve backwards compatibility with ASM 4.

While in theory, anyone using pojomatic could simply ensure that their project
uses ASM 6 or later, this commit removes that burden. Should we reach a point where
compatibility with ASM < 6 is no longer desired, then we can revert the changes
in this commit.