Add Java Platform Module System support (github issue 13).
Compile all code under java 7, except for module-info.java and LookupClassDefiner.java. If building under Java 9 or later, also compile those two classes, but using Java 9.

When running in a modular environment, using ClassLoader's defineClass
fails. In its place, MethodHandles.Lookup received an additional method,
defineClass. If this method exists (i.e. we are running Java 9 or later),
then use it; else stick with the old definition method.

A new sub (maven) module, modular-test-project, is used to verify that
all of this works like its supposed to. This is only run if using Java 9
or later to build.