Fix Java 11 issue with nested classes (github issue 14)
When scanning classes to determine property order, if the class is:
  a) compiled with JDK 11 or later, and
  b) has a nested class,
then ASM will refuse to scan the class unless the visitor declares itself
to be compliant with ASM version 7 or later.

Hence, while previously we went with ASM 6 to minimize impact to projects
that wished to not go to ASM 7, its clear now that it is better to go
with ASM 7. In theory, at least, this shouldn't break other projects,
as ASM is supposed to be backwards compatible these-a-days.

Testing this is a bit tricky, since we compile our tests targeting
JDK 1.7. Hence, to test this, we need to generate JDK-11 bytecode. We
use the Java Compiler API to do this.

Thanks to Bulat Nigmatullin for reporting this issue.